only send missing blocks if the last max is not equal to the new max
